+++ apps/api/src/core/repositories/base.repository.ts
@@ -1,6 +1,6 @@
 import type { AnyAbility } from "@casl/ability";
 import type { DBQueryConfig } from "drizzle-orm";
-import { and, eq, inArray, sql } from "drizzle-orm";
+import { and, eq, inArray, isNull, sql } from "drizzle-orm";
 import type { PgTableWithColumns } from "drizzle-orm/pg-core";
 import type { SQL } from "drizzle-orm/sql/sql";
 
@@ -192,8 +192,10 @@ export abstract class BaseRepository<
   protected queryToWhere(query: Partial<T["$inferSelect"]> | undefined) {
     if (!query) return this.whereAccessibleBy(undefined);
 
-    const fields = Object.keys(query) as Array<keyof T["$inferSelect"]>;
-    const where = fields.length ? and(...fields.map(field => eq(this.table[field], query[field]))) : undefined;
+    const fields = Object.keys(query).filter(key => query[key as keyof typeof query] !== undefined) as Array<keyof T["$inferSelect"]>;
+    const where = fields.length
+      ? and(...fields.map(field => (query[field] === null ? isNull(this.table[field]) : eq(this.table[field], query[field]))))
+      : undefined;
 
     return this.whereAccessibleBy(where);
   }